Bump Klarna tests
