+++ Example/PrimerSDKExample_UITests/Base.swift
@@ -71,21 +71,5 @@ class Base: XCTestCase {
         app.tables.buttons.containing(.staticText, identifier:"•••• 1111").firstMatch.tap()
         app.alerts["Confirmation"].scrollViews.otherElements.buttons["Delete"].tap()
     }
-    
-    func testInitKlarna() throws {
-        try testPresentWallet()
-
-        app.buttons["vault_klarna_button"].tap()
-
-        let exists = NSPredicate(format: "exists == 1")
-        
-        let webView = app.webViews["primer_webview"]
-        expectation(for: exists, evaluatedWith: webView, handler: nil)
-        waitForExpectations(timeout: 5, handler: nil)
-        
-        let continueButton = app.webViews.buttons["Continue"]
-        expectation(for: exists, evaluatedWith: continueButton, handler: nil)
-        waitForExpectations(timeout: 30, handler: nil)
-    }
 
 }
